Began integration of the CPU emulation in with the video emulation Implemented the CPU emulation library Added a documentation folder
CZX81 now functionally emulates the Z80 CPU and memory space!
at debug level 2, you can see the disassembly of each executed instruction

The video emulation handles the display of memory wrong as of now. Currently it treats the video memory as a linear charbuffer
It needs to be re-written to instead treat the video memory as a file (DFILE emulation)